+++ src/components/Posts/Cards/ImportantPostPreviewCard/ImportantPostPreviewCard.tsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router-dom';
 import Typography from '@material-ui/core/Typography';
 import { useStyles } from './ImportantPostPreviewCard.styles';
 import { IPostPreviewCardProps } from '../types';
-import background from '../../../../assets/images/mock_img_slider_bg.png';
+import background from '../mock_img_slider_bg.png';
 
 export const ImportantPostPreviewCard: React.FC<IPostPreviewCardProps> = (
   props,
